requirements.txt
flask
requests
openai
pinecone-client
flask-cors
python-dotenv
gunicorn
sseclient-py
langchain
PyPDF2